Move metadata to the viewlet template
