+++ sc/social/like/browser/viewlets.py
@@ -1,11 +1,17 @@
 # -*- coding: utf-8 -*-
+from Acquisition import aq_inner
+from Acquisition import aq_parent
 from plone import api
 from plone.api.exc import InvalidParameterError
 from plone.app.layout.viewlets import ViewletBase
 from plone.memoize.view import memoize
+from Products.CMFCore.interfaces import ISiteRoot
+from Products.CMFCore.utils import getToolByName
 from Products.CMFCore.WorkflowCore import WorkflowException
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from sc.social.like.behaviors import ISocialMedia
 from sc.social.like.interfaces import ISocialLikeSettings
+from sc.social.like.utils import get_content_image
 from zope.component import getMultiAdapter
 
 
@@ -61,7 +67,16 @@ class SocialMetadataViewlet(BaseLikeViewlet):
     """Viewlet used to insert metadata into page header
     """
     render = ViewPageTemplateFile('templates/metadata.pt')
-    render_method = 'metadata'
+
+    def __init__(self, context, request, view, manager):
+        super(SocialMetadataViewlet, self).__init__(context, request, view, manager)
+        self.title = context.title
+        self.description = context.Description()
+        self.portal_state = getMultiAdapter((self.context, self.request),
+                                            name=u'plone_portal_state')
+        self.site_url = self.portal_state.portal_url()
+        self.url = context.absolute_url()
+        self.image = get_content_image(context)
 
     def enabled(self):
         """Validates if the viewlet should be enabled for this context
@@ -73,6 +88,49 @@ class SocialMetadataViewlet(BaseLikeViewlet):
         if template in ('all_content', 'folder_full_view',):
             return True
         return self.helper.enabled(self.view)
+
+    @property
+    def via(self):
+        record = ISocialLikeSettings.__identifier__ + '.twitter_username'
+        try:
+            return api.portal.get_registry_record(record)
+        except InvalidParameterError:
+            return ''
+
+    @property
+    def canonical_url(self):
+        if ISocialMedia.providedBy(self.context):
+            return self.context.canonical_url
+        else:
+            # use current URL if the object don't provide the behavior
+            return self.url
+
+    def image_url(self):
+        """ Return url to image
+        """
+        img = self.image
+        if img:
+            return img.url
+        else:
+            return '{0}/logo.png'.format(self.site_url)
+
+    def _isPortalDefaultView(self):
+        context = self.context
+        if ISiteRoot.providedBy(aq_parent(aq_inner(context))):
+            putils = getToolByName(context, 'plone_utils')
+            return putils.isDefaultPage(context)
+        return False
+
+    def _isPortal(self):
+        context = self.context
+        if ISiteRoot.providedBy(aq_inner(context)):
+            return True
+        return self._isPortalDefaultView()
+
+    def type(self):
+        if self._isPortal():
+            return 'website'
+        return 'article'
 
 
 class SocialLikesViewlet(BaseLikeViewlet):
