+++ sc/social/like/plugins/email/browser.py
@@ -11,7 +11,6 @@ class PluginView(BrowserView):
     typebutton = ''
     language = 'en'
 
-    metadata = None
     plugin = link = ViewPageTemplateFile('plugin.pt')
 
     @property
