+++ sc/social/like/plugins/facebook/browser.py
@@ -1,17 +1,11 @@
 # -*- coding:utf-8 -*-
-from Acquisition import aq_inner
-from Acquisition import aq_parent
 from plone import api
 from plone.api.exc import InvalidParameterError
-from Products.CMFCore.interfaces import ISiteRoot
-from Products.CMFCore.utils import getToolByName
 from Products.Five import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from sc.social.like.behaviors import ISocialMedia
-from sc.social.like.config import IS_PLONE_5
 from sc.social.like.interfaces import ISocialLikeSettings
 from sc.social.like.plugins.facebook.utils import facebook_language
-from sc.social.like.utils import get_content_image
 from sc.social.like.utils import get_language
 from urllib import urlencode
 from zope.component import getMultiAdapter
@@ -26,7 +20,6 @@ class PluginView(BrowserView):
     fb_enabled = False
     language = 'en_US'
 
-    metadata = ViewPageTemplateFile('templates/metadata.pt')
     plugin = ViewPageTemplateFile('templates/plugin.pt')
     link = ViewPageTemplateFile('templates/link.pt')
 
@@ -40,12 +33,9 @@ class PluginView(BrowserView):
         self.description = context.Description()
         self.portal_state = getMultiAdapter((self.context, self.request),
                                             name=u'plone_portal_state')
-        self.portal = self.portal_state.portal()
         self.site_url = self.portal_state.portal_url()
-        self.portal_title = self.portal_state.portal_title()
         self.url = context.absolute_url()
         self.language = facebook_language(get_language(context), self.language)
-        self.image = get_content_image(context, width=1200, height=630)
         self.typebutton  # XXX: needed to initialize self.width
 
     @property
@@ -55,10 +45,6 @@ class PluginView(BrowserView):
         else:
             # use current URL if the object don't provide the behavior
             return self.url
-
-    @property
-    def is_plone_5(self):
-        return IS_PLONE_5
 
     def fbjs(self):
         js_source = """
@@ -71,37 +57,6 @@ class PluginView(BrowserView):
     }}());
     """.format(self.language)
         return js_source
-
-    def image_height(self):
-        """ Return height to image
-        """
-        img = self.image
-        if img:
-            return img.height
-
-    def image_type(self):
-        """ Return content type to image
-        """
-        img = self.image
-        if img:
-            return getattr(img, 'content_type',
-                           getattr(img, 'mimetype', 'image/jpeg'))
-
-    def image_width(self):
-        """ Return width to image
-        """
-        img = self.image
-        if img:
-            return img.width
-
-    def image_url(self):
-        """ Return url to image
-        """
-        img = self.image
-        if img:
-            return img.url
-        else:
-            return '{0}/logo.png'.format(self.site_url)
 
     @property
     def typebutton(self):
@@ -166,24 +121,6 @@ class PluginView(BrowserView):
             return 'Share' in api.portal.get_registry_record(record)
         except InvalidParameterError:
             return False
-
-    def _isPortalDefaultView(self):
-        context = self.context
-        if ISiteRoot.providedBy(aq_parent(aq_inner(context))):
-            putils = getToolByName(context, 'plone_utils')
-            return putils.isDefaultPage(context)
-        return False
-
-    def _isPortal(self):
-        context = self.context
-        if ISiteRoot.providedBy(aq_inner(context)):
-            return True
-        return self._isPortalDefaultView()
-
-    def type(self):
-        if self._isPortal():
-            return 'website'
-        return 'article'
 
     def share_link(self):
         params = dict(
