--- sc/social/like/plugins/facebook/templates/metadata.pt
@@ -1,20 +1,0 @@
-<tal:comment replace="nothing">
-  Plone 5 already includes metadata for Facebook and Twitter.
-</tal:comment>
-<tal:fb condition="not:view/is_plone_5">
-  <meta property="og:site_name" tal:attributes="content view/portal_title" />
-  <meta property="og:url" tal:attributes="content view/canonical_url" />
-  <meta property="og:type" tal:attributes="content view/type" />
-  <meta property="og:locale" tal:attributes="content view/language" />
-  <meta property="og:title" tal:attributes="content view/title" />
-  <meta property="og:description" tal:attributes="content view/description" />
-  <meta property="og:image" tal:attributes="content view/image_url" />
-  <tal:image tal:condition="nocall:view/image">
-  <meta property="og:image:height" tal:attributes="content view/image_height" />
-  <meta property="og:image:width" tal:attributes="content view/image_width" />
-  <meta property="og:image:type" tal:attributes="content view/image_type" />
-  </tal:image>
-  <meta tal:condition="view/admins" property="fb:admins" tal:attributes="content view/admins" />
-  <meta tal:condition="view/app_id" property="fb:app_id" tal:attributes="content view/app_id" />
-  <script type="application/javascript" tal:content="view/fbjs"></script>
-</tal:fb>
