+++ sc/social/like/plugins/facebook/templates/plugin.pt
@@ -1,16 +1,15 @@
-<tal:fb>
-    <!-- Facebook -->
-    <div class="fb-like"
-         tal:condition="view/fbshow_like"
-         tal:attributes="data-href view/canonical_url;
-                         data-send string:false;
-                         data-action view/fbaction;
-                         data-layout view/typebutton;
-                         data-width view/width;
-                         data-show-faces string:false;
-                         data-share view/fbshow_share"></div>
-    <div class="fb-share-button"
-         tal:condition="python: view.fbshow_share and not view.fbshow_like"
-         tal:attributes="data-layout view/typebutton;
-                         data-href view/canonical_url;"></div>
-</tal:fb>
+<!-- Facebook -->
+<script type="application/javascript" tal:content="view/fbjs" />
+<div class="fb-like"
+     tal:condition="view/fbshow_like"
+     tal:attributes="data-href view/canonical_url;
+                     data-send string:false;
+                     data-action view/fbaction;
+                     data-layout view/typebutton;
+                     data-width view/width;
+                     data-show-faces string:false;
+                     data-share view/fbshow_share"></div>
+<div class="fb-share-button"
+     tal:condition="python: view.fbshow_share and not view.fbshow_like"
+     tal:attributes="data-layout view/typebutton;
+                     data-href view/canonical_url;"></div>
