+++ sc/social/like/plugins/gplus/browser.py
@@ -14,7 +14,6 @@ class PluginView(BrowserView):
     gp_enabled = True
     language = 'en'
 
-    metadata = ViewPageTemplateFile('templates/metadata.pt')
     plugin = ViewPageTemplateFile('templates/plugin.pt')
     link = ViewPageTemplateFile('templates/link.pt')
