--- sc/social/like/plugins/gplus/templates/metadata.pt
@@ -1,12 +1,0 @@
-<tal:gp>
-  <script type="application/javascript">
-    (function() {
-        var po = document.createElement('script');
-        po.type = 'application/javascript';
-        po.async = true;
-        po.src = '//apis.google.com/js/plusone.js';
-        var head = document.getElementsByTagName('head')[0];
-        head.appendChild(po);
-    })();
-  </script>
-</tal:gp>
