--- sc/social/like/plugins/linkedin/templates/metadata.pt
@@ -1,13 +1,0 @@
-<tal:linkedin>
-    <script type="application/javascript">
-        (function() {
-            var po = document.createElement('script');
-            po.type = 'application/javascript';
-            po.async = true;
-            po.src = '//platform.linkedin.com/in.js?async=true';
-            po.onload = function() {IN.init()};
-            var head = document.getElementsByTagName('head')[0];
-            head.appendChild(po);
-        })();
-    </script>
-</tal:linkedin>
