--- sc/social/like/plugins/pinterest/templates/metadata.pt
@@ -1,3 +1,0 @@
-<tal:pinterest>
-
-</tal:pinterest>
