+++ sc/social/like/plugins/twitter/browser.py
@@ -5,9 +5,7 @@ from Products.CMFPlone.utils import safe_unicode
 from Products.Five import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.PythonScripts.standard import url_quote
-from sc.social.like.config import IS_PLONE_5
 from sc.social.like.interfaces import ISocialLikeSettings
-from sc.social.like.utils import get_content_image
 from sc.social.like.utils import get_language
 from urllib import urlencode
 from zope.component import getMultiAdapter
@@ -18,7 +16,6 @@ class PluginView(BrowserView):
     twitter_enabled = False
     language = 'en'
 
-    metadata = ViewPageTemplateFile('templates/metadata.pt')
     plugin = ViewPageTemplateFile('templates/plugin.pt')
     link = ViewPageTemplateFile('templates/link.pt')
 
@@ -37,7 +34,6 @@ class PluginView(BrowserView):
         self.portal_title = self.portal_state.portal_title()
         self.url = context.absolute_url()
         self.language = get_language(context)
-        self.image = get_content_image(context)
         self.urlnoscript = (
             u'http://twitter.com/home?status=' +
             url_quote(u'{0} - {1} via {2}'.format(
@@ -46,10 +42,6 @@ class PluginView(BrowserView):
                 self.via)
             )
         )
-
-    @property
-    def is_plone_5(self):
-        return IS_PLONE_5
 
     @property
     def typebutton(self):
@@ -77,7 +69,3 @@ class PluginView(BrowserView):
 
         url = 'https://twitter.com/intent/tweet?' + urlencode(params)
         return url
-
-    def image_url(self):
-        """Return image URL."""
-        return self.image.url if self.image else None
